Move PriorFactor.h to gtsam/nonlinear. Point slam/PriorFactor.h to new header location for backwards compatibility.
